apps/backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[project]
name = "pms-backend"
version = "0.1.0"
description = "PMS Backend API Service"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "motor>=3.6.0",
    "beanie>=1.27.0",
    "redis[asyncio]>=5.2.0",
    "aio-pika>=9.5.0",
    "opensearch-py[async]>=2.7.0",
    "aioboto3>=13.2.0",
    "pydantic-settings>=2.6.0",
    "structlog>=24.4.0",
    "prometheus-client>=0.21.0",
    "opentelemetry-sdk>=1.28.0",
    "opentelemetry-instrumentation-fastapi>=0.49b0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.12",
    "passlib[argon2]",
    "pyotp>=2.9.0",
    "qrcode[pil]>=7.4.2",
    "email-validator>=2.2.0",
    "reportlab>=4.2.0",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.27.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "fakeredis>=2.26.0",
    "mongomock-motor>=0.0.34",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "argon2-cffi",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = [
    "ignore::DeprecationWarning:passlib",
    "ignore::DeprecationWarning:argon2",
]
# Custom CLI flags (see tests/conftest.py)
# --keep-data   skip post-test teardown so data persists for inspection
# --real-db     use real MongoDB (MONGODB_URL env var) instead of mongomock

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
strict = false
ignore_missing_imports = true
